% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{theme_pub}
\alias{theme_pub}
\title{A ggplot2 theme for publication-quality graphics}
\usage{
theme_pub(base_size = 12)
}
\arguments{
\item{base_size}{Numeric. Base font size for the theme. Default is 12.}
}
\value{
A ggplot2 theme object.
}
\description{
\code{theme_pub()} is a custom ggplot2 theme designed for creating publication-quality
graphics with a clean and professional appearance. It builds upon the
\code{theme_minimal()} base theme and customizes various elements to enhance
readability and aesthetics.
}
