% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{years_available}
\alias{years_available}
\title{Get the range of years available for each dataset}
\usage{
years_available()
}
\value{
A named list with the range of years available for each dataset
}
\description{
Get the range of years available for each dataset
}
